Refactoring du transpose et documentation

+++ src/soundApi/soundApi.h
@@ -58,13 +58,13 @@ void snd_init_kits();
  * \param soundlength Durée du son 0 à 3f (attention valeur inversée: 3f = court) ACTIF seulement si loopmode = 1
  * \param loopmode Timed 1, continuous 0
  * \param output sortie du son 0 PAS DE SON / 1 GAUCHE / 2 DROITE / 3 GAUCHE ET DROITE
- * \param sfreq Le numéro de fréquence de la note jouée. de 0 à 72 cf: const u16 freqs[NB_FREQUENCES]
+ * \param sfreq Le numéro de fréquence de la note jouée. de 0 à 72 cf: const u8 freqs[NB_FREQUENCES]
  * \param transpose La valeur de transposition de 0 à FF
  **/
 void snd_playSoundOnChannel1(
         u16 sweeptime, u16 sweepdir, u16 sweepshifts, u16 volume,
         u16 envdir, u16 envsteptime, u16 waveduty, u16 soundlength,
-        u16 loopmode, u8 output, u16 sfreq, u8 transpose);
+        u16 loopmode, u8 output, u8 sfreq, u8 transpose);
 
 /**
  * \brief Joue un son sur le channel 1.
@@ -73,7 +73,7 @@ void snd_playSoundOnChannel1(
  * \param envelope Combinaison de tous les paramètres pour l'enveloppe (volume, envdir, envsteptime, waveduty) de 0 à FFFF
  * \param mode Timed 1, continuous 0
  * \param length Durée du son 0 à 3f (attention valeur inversée: 3f = court) ACTIF seulement si loopmode = 1
- * \param noteFreq Le numéro de fréquence de la note jouée. de 0 à 72 cf: const u16 freqs[NB_FREQUENCES]
+ * \param noteFreq Le numéro de fréquence de la note jouée. de 0 à 72 cf: const u8 freqs[NB_FREQUENCES]
  * \param transpose La valeur de transposition de 0 à FF.
  * 
  * Cette méthode est plus simple que snd_playSoundOnChannel1 mais produit le même résultat. 
@@ -91,12 +91,12 @@ void snd_simple_playSoundOnChannel1(u8 sweep, u16 envelope, u8 mode,
  * @param soundlength durée du son 0 à 3f (attention valeur inversée: 3f = court) ACTIF seulement si loopmode = 1
  * @param loopmode continuous 0, timed 1
  * @param output sortie du son 0 PAS DE SON / 1 GAUCHE / 2 DROITE / 3 GAUCHE ET DROITE
- * @param sfreq le numéro de fréquence de la note jouée. de 0 à 72 cf: const u16 freqs[NB_FREQUENCES]
+ * @param sfreq le numéro de fréquence de la note jouée. de 0 à 72 cf: const u8 freqs[NB_FREQUENCES]
  * @param transpose le paramètre de transposition
  */
 void snd_playSoundOnChannel2(u16 volume,
         u16 envdir, u16 envsteptime, u16 waveduty, u16 soundlength,
-        u16 loopmode, u8 output, u16 sfreq, u8 transpose);
+        u16 loopmode, u8 output, u8 sfreq, u8 transpose);
 
 /**
  * \brief Méthode simplifiée pour jouer un son sur le channel 2. <b>NON IMPLEMENTE!</b>
@@ -124,7 +124,7 @@ void snd_simple_playSoundOnChannel2(u16 envelope,
  * @param transpose le paramètre de transposition
  */
 void snd_playSoundOnChannel3(u16 volume, u16 soundLength, u16 loopmode, u16 voice,
-        u16 bank, u16 bankMode, u8 output, u16 freq, u8 transpose);
+        u16 bank, u16 bankMode, u8 output, u8 freq, u8 transpose);
 
 /**
  * \brief Méthode simplifiée pour jouer un son sur le channel 3. <b>NON IMPLEMENTE!</b>
@@ -146,7 +146,7 @@ void snd_simple_playSoundOnChannel3();
  * @param transpose
  */
 void snd_playSoundOnChannel4(u16 volume, u16 envdir, u16 envsteptime, u16 soundlength,
-        u16 loopmode, u8 output, u16 shiftFreq, u16 stepping, u16 freqRatio, u8 transpose);
+        u16 loopmode, u8 output, u16 shiftFreq, u16 stepping, u8 freqRatio, u8 transpose);
 
 /**
  * \brief Méthode simplifiée pour jouer un son sur le channel 4. <b>NON IMPLEMENTE!</b>
